Extracting animation out of FadeInImage
